network fee fix and add try catch to address handler

+++ trustscore/src/main/java/io/coti/trustscore/config/rules/UserNetworkFeeByTrustScoreRange.java
@@ -3,13 +3,15 @@ package io.coti.trustscore.config.rules;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class UserNetworkFeeByTrustScoreRange {
 
     private String type;
     private double limit;
-    private double minRate;
-    private double maxRate;
-    private double feeRate;
+    private BigDecimal minRate;
+    private BigDecimal maxRate;
+    private BigDecimal feeRate;
 }
